Backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
bcrypt
PyJWT
python-dotenv
pydantic
